Week_7/pyproject.toml
[project]
name = "week07-doc-intel"
version = "0.1.0"
description = "Standalone AI Document Intelligence Service (upload -> structured extraction)"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "pydantic>=2.6",
  "python-multipart>=0.0.9",
  "pypdf>=4.0.0",
  "python-docx>=1.1.0",
  "openai>=1.40.0"
]